Travis-CI: Optimize rate of false negatives vs true failures
Move 'encryption' tests to another job, since the 'binlog' and 'rpl'
tests are so slow and often make the job timeout (after 50 minutes).

Allow failure in ppc64el as it frequently fails due to out-of-memory.
A simple restart often fixes it, but we can't depend on restarts.

Also re-enable arm64 as MDEV-23955 is now fixed.

MERGING: This commit is OK to merge 10.6 and upwards.