MDEV-23072 Diskspace not reused for Blob in data file
- This issue is caused by commit a4948dafcd7eee65f16d848bdc6562fc49ef8916.
Purge doesn't free the externally stored page associated with the
last record of the root page. In that case, purge thread does empty
the root page and leads to more orphaned blob page in the tablespace.
Purge thread should free the blob even for the last record of the
root page.

Reviewed-by: Marko Mäkelä