MDEV-22974: mysql_native_password make '*40chars' invalid if nonhex
Porting a show create user from mysql may have the hashed password
string: *THISISNOTAVALIDPASSWORDTHATCANBEUSEDHERE.

Obviously this isn't a hash because it contains non-hex characters.

We do however recognise the *{40 chars not all are hex} as a pattern
for an invalid password and processes that internally. This was
determined to be the general pattern that MySQL used.

We use use same trick with "invalid" so that someone doesn't work out
what this hash this invalid string does correspond to (because
non-hex characters are accepted by the get_salt_from_password
password).

As a consequence of being a little more lax, we've also made it strict
that the 41 character has to begin with a '*'.