MDEV-24015: SQL Error (1038): Out of sort memory when enough memory for the sort buffer is provided
The issue here was with the re-execution of a subquey, after few executions
the subquery execution returned error that not enough space was provided inside the sort
buffer.
This is happening because with the introduction of packed sort keys, during the phase
where sort length is calculated the parameters of the sort keys are not reset and
we keep up adding the sort length to the previous calculated sort length for the
earlier executions of the subquery. This happens because the sort keys structure is
allocated only once and so if their is re-execution of a subquery involving sorting
then we should be resetting the parameters of the sort keys.