MDEV-24024 innodb.ibuf_not_empty failed in buildbot
Probably due to the changes to page flushing in MDEV-23399
(commit 7cffb5f6e8a231a041152447be8980ce35d2c9b8) the command
CHECK TABLE would occasionally report a different number of rows
for the corrupted secondary index. (The reported number was 991
instead of 990 on one occasion.)

Let us map all numbers to 990 in the output. We only care that the
injected corruption will be detected.