MDEV-24190 Accessing INFORMATION_SCHEMA.INNODB_TABLESPACES_ENCRYPTION may break innodb_open_files logic
If the function i_s_tablespaces_encryption_fill_table() was not able to
report all content to the SQL layer, it would fail to decrement
fil_system.freeze_space_list that it had incremented. This would
prevent the not-frequently-used logic from working, potentially causing
frequently used files to be closed and reopened whenever innodb_open_files
is exceeded.

This regression was caused by
commit 45ed9dd957eebc7fc84feb2509f4aa6baa908a95 (part of MDEV-23855).