MDEV-18976 Implement a CHECKSUM redo log record for improved validation
New OPTION CHECKSUM redo record is added. The records are generated on
mtr_t::commit() for each modified page, checksum is counted for the page and
added in redo log record. When the record is applied, the checksum is
compared to the counted page checksum, and warning is issued in error log if
it does not match.

We can't use full crc32 algoritms for checksum counting because:

a) FIL_PAGE_FILE_FLUSH_LSN_OR_KEY_VERSION field can be changed outside
of minitransaction during page flushing. see stack trace:
▾ fil_encrypt_buf_for_non_full_checksum
  ▾ fil_encrypt_buf
      ▾ fil_space_encrypt
            ▾ buf_page_encrypt
                    ▾ buf_flush_page

b) the last 4 bytes of FIL_PAGE_LSN is firstly written to the page at
srv_page_size-FIL_PAGE_FCRC32_END_LSN offset, and then checksum is
counted for the whole page including just written value. But
FIL_PAGE_LSN is updated after mtr commit.

That is why buf_calc_page_crc32() is used.

CHECkSUM records are not written for ROW_FORMAT=COMPRESSED pages to
simplify the code.

The new option innodb_log_page_checksum is added, the option turns on the above
functionality.

The records are not generated for freed pages within the mtr. For this
puprose the code for pages scrubbing is reused. It collects freed page
ids within the mtr. The code was moved from buf_page_free() to
mtr_t::free() because there can be the cases when mtr_t::free() is
called without buf_page_free(), for example, when free segments are
purged in purge coordinator thread. The code is also added in
fseg_free_extent() because it invokes buf_page_free() without mtr_t::free().