MDEV-24448 srv_start(): Assertion !buf_pool.any_io_pending()
We are seeing !buf_pool.any_io_pending() assertion failures
in srv_start() ever since MDEV-21452 in 10.6. But, the
problem appears to be older. In 10.5 since MDEV-19514 removed
writes from the precursor of buf_page_read_complete(), it
seems that the debug assertion failure could have been harmless.

recv_sys_t::apply(): At the end of each batch, wait not only
for all log records to have been processed, but also for all
pending reads to complete, so that the buffer pool will be in
an idle state.