Deb: Refactor autopkgtests
Extend Debian autopkgtest to test for all InnoDB compression methods
that are expected to work and disable those that should not (e.g. Snappy).

This upstreams the changes of downstream Debian commit
https://salsa.debian.org/mariadb-team/mariadb-10.3/-/commit/278531a7dfa7d60a60b067d089860c92a4e1221b
apart from having Snappy support.

In addition, simplify Debian autopkgtest RocksDB part and make it more
verbose so that future regressions like this are easier to debug.

Snappy is still used in RocksDB and ColumnStore and autopkgtests thus
also check that it is available in RocksDB.